src/bot/text-to-cad/requirements.txt
kittycad
python-dotenv
gradio